Fix Error Caused By Empty Linker Options
The Error when building djinn:

    Linking dist/dist-sandbox-865448ff/build/djinn/djinn ...
    gcc: error: unrecognized command line option ‘-Wl’
    cabal: Error: some packages failed to install:
    djinn-2011.7.23 failed during the building phase. The exception was:
    ExitFailure 1

Caused by `-optionl-Wl` in ghc-options field inside djinn.cabal.  `-optionl`
for ghc is a command line options for passing it's argument to the
linker, in this case the linker is GCC.  This option wasn't recognized by
GCC version 4.7.3, since the -Wl options must followed by an comma separated
command line options that will be passed to the linker. In previous version,
especially GCC version 4.6.3, this empty -Wl option will be ignored, so
there are no problem installing on older version of GCC.